% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addShadedNight.R
\name{addShadedNight}
\alias{addShadedNight}
\title{Add nighttime shading to a timeseries plot}
\usage{
addShadedNight(timeInfo, col = adjustcolor("black", 0.1))
}
\arguments{
\item{timeInfo}{dataframe as returned by \code{monitor_timeInfo()}}

\item{col}{Color used to shade nights.}
}
\description{
Draw shading rectangles on a plot to indicate nighttime hours.
}
\seealso{
\link{timeInfo}
}
