% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirMonitor.R
\docType{data}
\name{AirFire_S3_archiveBaseUrl}
\alias{AirFire_S3_archiveBaseUrl}
\title{USFS maintained archive base URL}
\format{
A url
}
\usage{
AirFire_S3_archiveBaseUrl
}
\description{
The US Forest Service AirFire group maintains an archive of
processed monitoring data. The base URL for this archive is used as the
default in all \code{~_load()} functions.

\preformatted{
"https://airfire-data-exports.s3.us-west-2.amazonaws.com/monitoring/v2"
}
}
\details{
AirFire_S3_archiveBaseUrl
}
\keyword{datasets}
