% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_loadAnnual.R
\name{airnow_loadAnnual}
\alias{airnow_loadAnnual}
\title{Load annual AirNow monitoring data}
\usage{
airnow_loadAnnual(
  year = NULL,
  parameterName = "PM2.5",
 
    archiveBaseUrl = "https://airfire-data-exports.s3.us-west-2.amazonaws.com/monitoring/v2",
  archiveBaseDir = NULL,
  QC_negativeValues = c("zero", "na", "ignore")
)
}
\arguments{
\item{year}{Year [YYYY].}

\item{parameterName}{One of the EPA AQS criteria parameter names.}

\item{archiveBaseUrl}{Base URL for monitoring v2 data files.}

\item{archiveBaseDir}{Local base directory for monitoring v2 data files.}

\item{QC_negativeValues}{Type of QC to apply to negative values.}
}
\value{
A \emph{mts_monitor} object with AirNow data. (A list with
\code{meta} and \code{data} dataframes.)
}
\description{
Loads pre-generated .rda files containing hourly AirNow data.

If \code{archiveDataDir} is defined, data will be loaded from this local
archive. Otherwise, data will be loaded from the monitoring data repository
maintained by the USFS AirFire team.

The files loaded by this function contain a single year's worth of data

For the most recent data in the last 10 days, use \code{airnow_loadLatest()}.

For daily updates covering the most recent 45 days, use \code{airnow_loadDaily()}.

For archival data for a specific month, use \code{airnow_loadMonthly()}.

Pre-processed AirNow exists for the following parameters:
\enumerate{
\item{PM2.5}
}
}
