% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirMonitor.R
\docType{data}
\name{CONUS}
\alias{CONUS}
\title{CONUS state codes}
\format{
A vector with 49 elements
}
\usage{
CONUS
}
\description{
State codes for the 48 contiguous states +DC that make up the CONtinental US.
}
\details{
CONUS state codes
}
\keyword{datasets}
