% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirMonitor-package.R
\docType{data}
\name{coreMetadataNames}
\alias{coreMetadataNames}
\title{Names of standard metadata columns}
\format{
A vector of character strings
}
\usage{
coreMetadataNames
}
\description{
Vector of names of the required \code{monitor$meta} columns.
These represent metadata columns that must exist in every valid
\emph{mts_monitor} object. Any number of additional columns may also be present.
}
\details{
coreMetadataNames
}
\examples{
 print(coreMetadataNames, width = 80)
}
\keyword{datasets}
