% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_aqs_loadAnnual.R
\name{epa_aqs_loadAnnual}
\alias{epa_aqs_loadAnnual}
\title{Load annual EPA AQS monitoring data}
\usage{
epa_aqs_loadAnnual(
  year = NULL,
  parameterCode = NULL,
  baseUrl = NULL,
  baseDir = NULL
)
}
\arguments{
\item{year}{Desired year (integer or character representing YYYY).}

\item{parameterCode}{EPA AQS Parameter Code.}

\item{baseUrl}{Base URL for annual EPA AQS data files.}

\item{baseDir}{Local base directory for annual EPA AQS data files.}
}
\value{
A \emph{mts_monitor} object with EPA AQS data.
}
\description{
Loads pre-generated .RData files containing annual
EPA data.

EPA parameter codes include:
\enumerate{
\item{42101}{ -- CO}
\item{88101}{ -- PM2.5 FRM/FEM Mass (begins in 2008)}
\item{88502}{ -- PM2.5 non FRM/FEM Mass (begins in 1998)}
}
}
