% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_bestTimezone.R
\name{monitor_bestTimezone}
\alias{monitor_bestTimezone}
\title{Return the most common timezone}
\usage{
monitor_bestTimezone(monitor = NULL)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object.}
}
\value{
A valid \code{base::OlsonNames()} timezone.
}
\description{
Evaluates all timezones in \code{monitor} and returns the
most common one. In the case of a tie, the alphabetically first one is
returned.
}
