% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-monitor.R
\name{monitor_check}
\alias{monitor_check}
\title{Check an \emph{monitor} object for validity.}
\usage{
monitor_check(monitor)
}
\arguments{
\item{monitor}{\emph{monitor} objet.}
}
\description{
Checks on the validity of an \emph{monitor} object. If any test
fails, this function will stop with a warning message.
}
