% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-monitor.R
\name{monitor_extractDataFrame}
\alias{monitor_extractDataFrame}
\alias{monitor_extractData}
\alias{monitor_extractMeta}
\title{Extract dataframes from \emph{monitor} objects}
\usage{
monitor_extractData(monitor)

monitor_extractMeta(monitor)
}
\arguments{
\item{monitor}{\emph{monitor} object to extract dataframe from.}
}
\value{
A dataframe from the given \emph{monitor} object
}
\description{
These functions are convenient wrappers for extracting the dataframes that
comprise a \emph{monitor} object. These functions are designed to be useful when
manipulating data in a pipeline chain using \code{\%>\%}.

Below is a table showing equivalent operations for each function.

\tabular{ll}{
  \strong{Function} \tab \strong{Equivalent Operation}\cr
  \code{monitor_extractData(monitor)} \tab \code{monitor$data}\cr
  \code{monitor_extractMeta(monitor)} \tab \code{monitor$meta}
}
}
