% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_fromPWFSLSmoke.R
\name{monitor_fromPWFSLSmoke}
\alias{monitor_fromPWFSLSmoke}
\title{Convert a ws_monitor object from the PWFSLSmoke package}
\usage{
monitor_fromPWFSLSmoke(ws_monitor = NULL)
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object. (A list with
\code{meta} and \code{data} dataframes.)}
}
\value{
A \emph{mts_monitor} object.
}
\description{
A \pkg{PWFSLSmoke} package \emph{ws_monitor} object is enhanced
and modified so that it becomes a valid \emph{mts_monitor} object. This is
a lossless operation and can be reversed with \code{monitor_toPWFSLSmoke()}.
}
