% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-monitor.R
\name{monitor_getDataFrame}
\alias{monitor_getDataFrame}
\alias{monitor_getData}
\alias{monitor_getMeta}
\title{Extract dataframes from \emph{mts_monitor} objects}
\usage{
monitor_getData(monitor)

monitor_getMeta(monitor)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object to extract dataframe from.}
}
\value{
A dataframe from the given \emph{mts_monitor} object.
}
\description{
These functions are convenient wrappers for extracting the dataframes that
comprise a \emph{mts_monitor} object. These functions are designed to be
useful when manipulating data in a pipeline using \code{\%>\%}.

Below is a table showing equivalent operations for each function.

\tabular{ll}{
  \strong{Function} \tab \strong{Equivalent Operation}\cr
  \code{monitor_getData(monitor)} \tab \code{monitor$data}\cr
  \code{monitor_getMeta(monitor)} \tab \code{monitor$meta}
}
}
