% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-monitor.R
\name{monitor_isEmpty}
\alias{monitor_isEmpty}
\title{Test for an empty \emph{mts_monitor} object}
\usage{
monitor_isEmpty(monitor)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object}
}
\value{
Invisibly returns \code{TRUE} if no data exist in \code{mts_monitor}, \code{FALSE} otherwise.
}
\description{
This function returns true under the following conditions:
\itemize{
\item{no time series: \code{ncol(monitor$data) == 1}}
\item{no time series records: \code{nrow(monitor$data) == 0}}
\item{all timeseries values are \code{NA}}
}
This makes for more readable code in functions that need to test for this.
}
