% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-monitor.R
\name{monitor_isValid}
\alias{monitor_isValid}
\title{Test \emph{mts_monitor} object for correct structure}
\usage{
monitor_isValid(monitor = NULL, verbose = FALSE)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object}

\item{verbose}{Logical specifying whether to produce detailed warning messages.}
}
\value{
Invisibly returns \code{TRUE} if \code{mts_monitor} has the correct
structure, \code{FALSE} otherwise.
}
\description{
The \code{mts_monitor} is checked for the presence of core
\code{meta} and \code{data} columns.

Core \code{meta} columns include: (TODO:  complete this list)

\itemize{
  \item{\code{deviceDeploymentID} -- unique identifier (see \pkg{MazmaLocationUtils})}
  \item{\code{deviceID} -- device identifier}
  \item{\code{locationID} -- location identifier (see \pkg{MazmaLocationUtils})}
  \item{\code{locationName} -- English language name}
  \item{\code{longitude} -- decimal degrees E}
  \item{\code{latitude} -- decimal degrees N}
  \item{\code{elevation} -- elevation of station in m}
  \item{\code{countryCode} -- ISO 3166-1 alpha-2}
  \item{\code{stateCode} -- ISO 3166-2 alpha-2}
  \item{\code{timezone} -- Olson time zone}
}

Core \code{data} columns include:

\itemize{
  \item{\code{datetime} -- measurement time (UTC)}
}
}
