% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_leaflet.R
\name{monitor_leaflet}
\alias{monitor_leaflet}
\title{Leaflet interactive map of monitor locations}
\usage{
monitor_leaflet(
  monitor,
  slice = get("max"),
  maptype = "terrain",
  extraVars = NULL,
  ...
)
}
\arguments{
\item{monitor}{\emph{monitor} object.}

\item{slice}{Either a time index or a function used to collapse the time axis
-- defautls to \code{get('max')}}

\item{maptype}{optional name of leaflet ProviderTiles to use, e.g. "terrain"}

\item{extraVars}{Character vector of addition \code{locationTbl} column names
to be shown in leaflet popups.}

\item{...}{Additional arguments passed to \code{leaflet::addCircleMarker()}.}
}
\value{
Invisbly returns a leaflet map of class "leaflet".
}
\description{
This function creates interactive maps that will be displayed in RStudio's
'Viewer' tab. The \code{slice} argument is used to collapse a
\emph{monitor} timeseries into a single value. If \code{slice} is an
integer, that row index will be selected from the \code{monitor$data}
dataframe. If \code{slice} is a function (unquoted), that function will be
applied to the timeseries with the argument \code{na.rm=TRUE} (e.g.
\code{max(..., na.rm=TRUE)}).

If \code{slice} is a user defined function it will be used with argument
\code{na.rm=TRUE} to collapse the time dimension. Thus, user defined
functions must accept \code{na.rm} as an argument.
}
\details{
The \code{maptype} argument is mapped onto leaflet "ProviderTile" names.
Current map types include:

\enumerate{
\item{"roadmap"}{ -- "OpenStreetMap"}
\item{"satellite"}{ -- "Esri.WorldImagery"}
\item{"terrain"}{ -- "Esri.WorldTopoMap"}
\item{"toner"}{ -- "Stamen.Toner"}
}

If a character string not listed above is provided, it will be used as the
underlying map tile if available. See
\url{https://leaflet-extras.github.io/leaflet-providers/} for a list of
"provider tiles" to use as the background map.
}
