% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_select.R
\name{monitor_select}
\alias{monitor_select}
\title{Select monitors within a \emph{mts_monitor} object}
\usage{
monitor_select(monitor, id)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object.}

\item{id}{Vector of \code{deviceDeploymentIDs}.}
}
\value{
A subset of the incoming \code{mts_monitor}.
}
\description{
Selects a subset of monitors within a \emph{mts_monitor} object.
This function basically performs \code{dplyr::select()} on \code{monitor$data}
and is more convenient than using \code{monitor_filterMeta()}. Compare;

\preformatted{
ids <- c(... some colleciton of deviceDeploymentIDs ...)

monitor %>%
  monitor_filter(deviceDeploymentID %in% ids)

monitor %>%
  monitor_select(ids)
}
}
\note{
Filtering is done on variables in \code{monitor$meta}.
}
\seealso{
\link{monitor_filterMeta}
}
