% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_toPWFSLSmoke.R
\name{monitor_toPWFSLSmoke}
\alias{monitor_toPWFSLSmoke}
\title{Convert a mts_monitor object to a ws_monitor object for the PWFSLSmoke package}
\usage{
monitor_toPWFSLSmoke(monitor = NULL)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object}
}
\value{
A \pkg{PWFSLSmoke} \emph{ws_monitor} object. (A list with
\code{meta} and \code{data} dataframes.)
}
\description{
A \emph{mts_monitor} object is modified so that it becomes
a \pkg{PWFSLSmoke} package \emph{ws_monitor} object. While some information
will be lost, this operation can be reversed with \code{monitor_fromPWFSLSmoke()}.
}
\note{
In order to avoid duplicated \code{monitorID} values in the returned
\emph{ws_monitor} object, the full \code{deviceDeploymentID} will be used
as the \code{monitorID}.
}
