% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_trimDate.R
\name{monitor_trimDate}
\alias{monitor_trimDate}
\title{Trim a \emph{monitor} object to full days}
\usage{
monitor_trimDate(monitor = NULL, timezone = NULL)
}
\arguments{
\item{monitor}{\emph{monitor} object.}

\item{timezone}{Olson timezone used to interpret dates.}
}
\value{
A subset of the given \emph{monitor} object.
}
\description{
Trims the date range of a \emph{monitor} object to local time date
boundaries which are \emph{within} the range of data. This has the effect
of removing partial-day data records at the start and end of the timeseries
and is useful when calculating full-day statistics.

Day boundaries are calculated using the specified \code{timezone} or, if
\code{NULL}, from \code{monitor$meta$timezone}.
}
