% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirMonitor.R
\docType{data}
\name{pollutantNames}
\alias{pollutantNames}
\title{Names of standard pollutants}
\format{
A vector of character strings
}
\usage{
pollutantNames
}
\description{
Character string identifiers of recognized pollutants.
}
\details{
pollutantNames
}
\keyword{datasets}
