% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_loadAnnual.R
\name{wrcc_loadAnnual}
\alias{wrcc_loadAnnual}
\title{Load annual WRCC monitoring data}
\usage{
wrcc_loadAnnual(
  year = NULL,
  archiveBaseUrl = NULL,
  archiveBaseDir = NULL,
  QC_negativeValues = c("zero", "na", "ignore")
)
}
\arguments{
\item{year}{Year [YYYY].}

\item{archiveBaseUrl}{Base URL for annual EPA AQS data files.}

\item{archiveBaseDir}{Local base directory for annual EPA AQS data files.}

\item{QC_negativeValues}{Type of QC to apply to negative values.}
}
\value{
A \emph{ws_monitor} object with WRCC data.
}
\description{
Loads pre-generated .rda files containing annual
WRCC data.

If \code{archiveDataDir} is defined, data will be loaded from this local
archive. Otherwise, data will be loaded from the monitoring data repository
maintained by the USFS AirFire team.

Current year files loaded by this function are updated once per week.

For the most recent data in the last 10 days, use \code{wrcc_loadLatest()}.

For daily updates covering the most recent 45 days, use \code{wrcc_loadDaily()}.
}
