% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAQILines.R
\name{addAQILines}
\alias{addAQILines}
\title{Add AQI lines to a plot}
\usage{
addAQILines(
  pollutant = c("PM2.5", "CO", "OZONE", "PM10", "AQI"),
  palette = c("EPA", "subdued", "deuteranopia"),
  ...
)
}
\arguments{
\item{pollutant}{EPA AQS criteria pollutant.}

\item{palette}{Named color palette to use for AQI categories.}

\item{...}{additional arguments to be passed to \code{abline()}}
}
\value{
No return value, called to add lines to a time series plot.
}
\description{
Draws AQI lines across a plot at the levels appropriate for
The \link{monitor_timeseriesPlot} function uses this function internally when
specifying \code{addAQI = TRUE}.
\code{pollutant}.
}
