% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addShadedNight.R
\name{addShadedNight}
\alias{addShadedNight}
\title{Add nighttime shading to a timeseries plot}
\usage{
addShadedNight(timeInfo, col = adjustcolor("black", 0.1))
}
\arguments{
\item{timeInfo}{dataframe as returned by \code{MazamaTimeSeries::monitor_timeInfo()}}

\item{col}{Color used to shade nights.}
}
\value{
No return value, called to add day/night shading to a timeseries plot.
}
\description{
Draw shading rectangles on a plot to indicate nighttime hours.
The \link{monitor_timeseriesPlot} function uses this function internally when
specifying \code{shadedNight = TRUE}.
}
