% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirMonitor.R
\docType{package}
\name{AirMonitor}
\alias{AirMonitor}
\alias{AirMonitor-package}
\title{Air Quality Data Analysis}
\description{
\code{
Utilities for working with hourly air quality monitoring data
with a focus on small particulates (PM2.5). A compact data model is
structured as a list with two dataframes. A 'meta' dataframe contains
spatial and measuring device metadata associated with deployments at known
locations. A 'data' dataframe contains a 'datetime' column followed by
columns of measurements associated with each "device-deployment".
}
}
