% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirMonitor-package.R
\docType{data}
\name{CONUS}
\alias{CONUS}
\title{CONUS state codes}
\format{
A vector with 49 elements
}
\usage{
CONUS
}
\description{
State codes for the 48 contiguous states +DC that make up the CONtinental US.

\code{
CONUS <- c(
  "AL","AZ","AR","CA","CO","CT","DE","FL","GA",
  "ID","IL","IN","IA","KS","KY","LA","ME","MD",
  "MA","MI","MN","MS","MO","MT","NE","NV","NH","NJ",
  "NM","NY","NC","ND","OH","OK","OR","PA","RI","SC",
  "SD","TN","TX","UT","VT","VA","WA","WV","WI","WY",
  "DC"
)
}
}
\details{
CONUS state codes
}
\keyword{datasets}
