% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-monitor.R
\name{monitor_check}
\alias{monitor_check}
\title{Check an \emph{mts_monitor} object for validity.}
\usage{
monitor_check(monitor)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object.}
}
\value{
Invisibly returns \code{TRUE} if \code{mts_monitor} has the correct
structure. Stops with an error message otherwise.
}
\description{
Checks on the validity of an \emph{mts_monitor} object. If any test
fails, this function will stop with a warning message.
}
