% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-monitor.R
\name{monitor_distinct}
\alias{monitor_distinct}
\title{Retain only distinct data records in monitor$data}
\usage{
monitor_distinct(monitor)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object}
}
\value{
A \emph{mts_monitor} object with no duplicated data records. (A list with
\code{meta} and \code{data} dataframes.)
}
\description{
Two successive steps are used to guarantee that the
\code{datetime} axis contains no repeated values:

\enumerate{
\item{remove any duplicate records}
\item{guarantee that rows are in \code{datetime} order}
}
}
\note{
This function is primarily for package-internal use.
}
