% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_dropEmpty.R
\name{monitor_dropEmpty}
\alias{monitor_dropEmpty}
\title{Drop device deployments with all missing data}
\usage{
monitor_dropEmpty(monitor)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object. (A list with
\code{meta} and \code{data} dataframes.)}
}
\value{
A subset of the incoming \code{mts_monitor}. (A list with
\code{meta} and \code{data} dataframes.)
}
\description{
The incoming \emph{mts_monitor} object is subset to retain
only time series with valid data.
}
