% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_filterMeta.R
\name{monitor_filterMeta}
\alias{monitor_filterMeta}
\alias{monitor_filter}
\title{General purpose metadata filtering for \emph{mts_monitor} objects}
\usage{
monitor_filterMeta(monitor, ...)

monitor_filter(monitor, ...)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object.}

\item{...}{Logical predicates defined in terms of the variables in
\code{monitor$meta}.}
}
\value{
A subset of the incoming \code{mts_monitor}.
}
\description{
A generalized metadata filter for \emph{mts_monitor} objects to
choose rows/cases where conditions are true.  Multiple conditions are
combined with \code{&} or separated by a comma. Only rows where the condition
evaluates to TRUE are kept. Rows where the condition evaluates to \code{NA}
are dropped.

`monitor_filter()` is just a convenient alias for `monitor_filterMeta()`.
}
\note{
Filtering is done on variables in \code{monitor$meta}.
}
\seealso{
\link{monitor_filter}

\link{monitor_filterDate}

\link{monitor_filterDatetime}
}
