% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_timeseriesPlot.R
\name{monitor_timeseriesPlot}
\alias{monitor_timeseriesPlot}
\title{Create Timeseries Plot}
\usage{
monitor_timeseriesPlot(
  monitor,
  shadedNight = FALSE,
  add = FALSE,
  addAQI = TRUE,
  opacity = NULL,
  ...
)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object.}

\item{shadedNight}{Logical specifying whether to add nighttime shading.}

\item{add}{Logical specifying whether to add to the current plot.}

\item{addAQI}{Logical specifying whether to add AQI levels and legend.}

\item{opacity}{Opacity to use for points. By default, an opacity is chosen based
on the number of points so that trends are highlighted while outliers diminish
in visual importance as the number of points increases.}

\item{...}{Additional arguments to be passed to \code{graphics::plot.default()}.}
}
\description{
Creates a time series plot of data from a \emph{mts_monitor} object.
By default, points are plotted as semi-transparent squares. All data values
are plotted from all monitors found in the \emph{mts_monitor} object.

Reasonable defaults are chosen for annotations and plot characteristics.
Users can override any defaults by passing in parameters accepted by
\code{graphics::plot.default}.
}
