% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirMonitor-package.R
\docType{data}
\name{pollutantNames}
\alias{pollutantNames}
\title{Names of standard pollutants}
\format{
A vector of character strings
}
\usage{
pollutantNames
}
\description{
Character string identifiers of recognized pollutant names.
}
\details{
pollutantNames
}
\examples{
 print(coreMetadataNames, width = 80)
}
\keyword{datasets}
