% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirMonitor.R
\docType{data}
\name{US_52}
\alias{US_52}
\title{US state codes}
\format{
A vector with 52 elements
}
\usage{
US_52
}
\description{
State codes for the 50 states +DC +PR (Puerto Rico).

\code{
US_52 <- c(
  "AK","AL","AZ","AR","CA","CO","CT","DE","FL","GA",
  "HI","ID","IL","IN","IA","KS","KY","LA","ME","MD",
  "MA","MI","MN","MS","MO","MT","NE","NV","NH","NJ",
  "NM","NY","NC","ND","OH","OK","OR","PA","RI","SC",
  "SD","TN","TX","UT","VT","VA","WA","WV","WI","WY",
  "DC","PR"
)
}
}
\details{
US state codes
}
\keyword{datasets}
